requirements.txt
gunicorn
flask
flask-sqlalchemy
psycopg2
line-bot-sdk